+++ .storybook/main.js
@@ -1,0 +1,12 @@
+module.exports = {
+  stories: ['../src/**/*.stories.(js|mdx)'],
+  addons: [
+    '@storybook/addon-actions',
+    '@storybook/addon-links',
+    '@storybook/addon-docs',
+    '@storybook/addon-knobs/register',
+    '@storybook/addon-viewport/register',
+    '@dump247/storybook-state/register',
+  ],
+};
+